Move instrument_view tests out of neutron tests
