+++ python/tests/test_neutron.py
@@ -5,25 +5,7 @@
 import scipp as sc
 import scippneutron as scn
 import numpy as np
-
-
-def make_dataset_with_beamline():
-    positions = [[1, 0, 0], [0, 1, 0], [0, 0, 1], [-1, 0, 0]]
-    d = sc.Dataset(
-        data={'a': sc.Variable(dims=['position', 'tof'], values=np.random.rand(4, 9))},
-        coords={
-            'tof':
-            sc.Variable(dims=['tof'],
-                        values=np.arange(1000.0, 1010.0),
-                        unit=sc.units.us),
-            'position':
-            sc.vectors(dims=['position'], values=positions, unit=sc.units.m)
-        })
-
-    d.coords['source_position'] = sc.vector(value=np.array([0, 0, -10]),
-                                            unit=sc.units.m)
-    d.coords['sample_position'] = sc.vector(value=np.array([0, 0, 0]), unit=sc.units.m)
-    return d
+from .common import make_dataset_with_beamline
 
 
 def test_neutron_convert():
@@ -54,31 +36,3 @@ def test_neutron_beamline():
     assert sc.identical(scn.L1(d) + scn.L2(d), scn.Ltotal(d, scatter=True))
     assert two_theta.unit == sc.units.rad
     assert two_theta.dims == ['position']
-
-
-def test_neutron_instrument_view_3d():
-    d = make_dataset_with_beamline()
-    scn.instrument_view(d["a"])
-
-
-def test_neutron_instrument_view_with_dataset():
-    d = make_dataset_with_beamline()
-    d['b'] = sc.Variable(dims=['position', 'tof'], values=np.arange(36.).reshape(4, 9))
-    scn.instrument_view(d)
-
-
-def test_neutron_instrument_view_with_masks():
-    d = make_dataset_with_beamline()
-    x = np.transpose(d.coords['position'].values)[0, :]
-    d['a'].masks['amask'] = sc.Variable(dims=['position'],
-                                        values=np.less(np.abs(x), 0.5))
-    scn.instrument_view(d["a"])
-
-
-def test_neutron_instrument_view_with_cmap_args():
-    d = make_dataset_with_beamline()
-    scn.instrument_view(d["a"],
-                        vmin=0.001 * sc.units.one,
-                        vmax=5.0 * sc.units.one,
-                        cmap="magma",
-                        norm="log")
